requirements.txt
mecab-python3==0.7
